+++ breeze-text/src/main/kotlin/com/windea/breezeframework/text/extensions/eng/SingularToPluralExtensions.kt
@@ -1,0 +1,2 @@
+package com.windea.breezeframework.text.extensions.eng
+